+++ xtime/time_test.go
@@ -21,6 +21,8 @@ var (
 	_ xml.Unmarshaler     = (*xtime.Time)(nil)
 	_ xmlstream.Marshaler = xtime.Time{}
 	_ xmlstream.WriterTo  = xtime.Time{}
+	_ xml.MarshalerAttr   = xtime.Time{}
+	_ xml.UnmarshalerAttr = (*xtime.Time)(nil)
 )
 
 type tokenReadEncoder struct {
@@ -48,5 +50,31 @@ func TestRoundTrip(t *testing.T) {
 
 	if !serverTime.Equal(respTime) {
 		t.Errorf("wrong time: want=%v, got=%v", serverTime, respTime)
+	}
+}
+
+func TestAttrMarshal(t *testing.T) {
+	zeroTime := time.Time{}.Add(24 * time.Hour)
+	xt := xtime.Time{Time: zeroTime}
+	name := xml.Name{Space: "example.net", Local: "foo"}
+	attr, err := xt.MarshalXMLAttr(name)
+	if err != nil {
+		t.Fatalf("unexpected error marshaling attr: %v", err)
+	}
+	if attr.Name != name {
+		t.Fatalf("wrong name for attr: want=%v, got=%v", name, attr.Name)
+	}
+	const expected = "0001-01-02T00:00:00Z"
+	if attr.Value != expected {
+		t.Fatalf("wrong value for attr: want=%v, got=%v", expected, attr.Value)
+	}
+
+	newTime := &xtime.Time{}
+	err = newTime.UnmarshalXMLAttr(attr)
+	if err != nil {
+		t.Fatalf("unexpected error unmarshaling attr: %v", err)
+	}
+	if *newTime != xt {
+		t.Fatalf("times don't match: want=%v, got=%v", xt, newTime)
 	}
 }